Change benchmark calculation to cycles per byte.
As is standard in cryptography. Currently we are at 17 CPB.
High speed steam cyphers like ChaCha20 can reach 2 CPB when
highly optimized with special instruction set. So there is
room for improvement. But the current speed might already be
fast enough, since the limiting factor is not the encryption
speed for the server, but I/O and state management.