Applied simple game session initialization and removed some opcodes.
It seems that the opcodes in the game file are actually the "events" that
the game server of tera knows. So it seems the initial server you
connect is just a proxy, that forwards your packets to other servers.

Maybe it's some kind of "global" server and all the instances, match-
macking and PVP arenas are different server systems.

I would also not be supprosed is the actul server uses the Tera client
in some way or the other.