Remove useless async cocde and return to STD.
Async is only needed in the network section of the server.
I reverted the codebase back to blocking code for every
other part for easier programming and compatibility with
the ECS we will use (which isn't async and is highly parallized
anyhow).

Also moved from tokio to async_std easier understanding the
async API, since the tokio API diverges in some places.