Switch license from GPLv3 to AGPLv3.
Since people will run this software mainly as a network service for
other people, we need a license that keeps their changes open-source
for other people.

If people run this server as a private server, they need to open their
source code to their users. If they didn't change anything of the
source files (not the configuration and data files), they of course
can linke their users to the official repository with an added hint
of the used version.