Update the deserialization.
Strings and arrays have an odd shape. They are "interleaved"
and we need to actually follow their offsets and only progress the
"self.pos" after their header once read.

So we maybe need to safe the old position somehow and return
it for these complex types. Hopefully we can simplify this
in the serialization.