Switch async runtime to async_std.
This refactoring was done to simplify the model layer
as to not support both sync and async APIs. With async_st
we can use task::block_on() everywhere without spinning
up a new runtime. This makes it possible to also run async
code in sync with the global ECS.

This also reduces the library count and we don't need to use
the rust-postgres, which has some insane customs (like running
the sync code in an tokio runtime or needed an extra connection
object for a query).

Resulting with this decision, we now use sqlx which seems to be
a more saner DB abstraction layer. We don't use the query! macro,
since DB access for compilation is borderline insane and we can
do the same with proper integration test.

We also could change the warp web framework with a more sane
version, that might be more infant in it's development stage and
doesn't support all features yet (like web sockets or HTTP2), but
doesn't break my cognitive flow with warps insane functional design
and mad type system when declaring handlers / endpoints.

The only real drawback is that async_std doesn't has the very nice
select! macro that tokio implemented, but I found another way to
have feature parity non the less.