Implement proper login functionality.
Login is now properly working and is tested. The overall persistence
infrastructure is now implemented and working as intended.

This commit heavily reworks the error handling to easy the writing of
the ECS systems.

We switched to a combination of this-error and anyhow to easier handle
the different kinds of error.

Also fixed some logging issues we had.