Rework the event ECS event handling.
Outgoing events are now send directly from a system using the channel
that is available inside a connection component. The major benefit
from this is, that a system doesn't need to borrow ViewMut and that way
is much easier to parallelize.

This change touched many parts, so I had to do it before we start writing
many other systems. Another benefit is that some places in the codebase
got a much cleaner.