Add UserLocation model and repository.
The UserLocation needs to save a rotation value. Since we will work
with nalgebra, I have decided to save the rotation as a vector.

This may waste some database space (since rotations of players
are always z_axis() based), but makes working with nalgebra much
easier (A Rotation3::angle() doesn't provide the full value angle of
a rotation, so you would need to also use the axis() component, this
is pretty awkward, so we will just save the scaled axis this way).